+++ src/frameserver/terminal/default/st/tui.c
@@ -267,7 +267,7 @@ static void on_mouse_button(
 }
 
 static void on_key(struct tui_context* c, uint32_t symest,
-	uint8_t scancode, uint8_t mods, uint16_t subid, void* t)
+	uint8_t scancode, uint16_t mods, uint16_t subid, void* t)
 {
 	char *str = NULL;
 	if (mods & TUIM_SHIFT){
